backend/motivation_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pyjwt>=2.1.0
python-multipart>=0.0.5
asyncpg>=0.25.0
pydantic-settings>=2.0.0
httpx==0.28.1
requests>=2.25.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
hyperframe>=6.0.1
h2>=4.0.0 